Agregando Policies

+++ app/Http/Controllers/Admin/PostController.php
@@ -87,6 +87,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        // Regla de Autorizacion
+        $this->authorize('author', $post);
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
         return view('admin.posts.edit', compact('post', 'categories', 'tags'));
@@ -101,6 +103,9 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
+        // Regla de Autorizacion
+        $this->authorize('author', $post);
+
         $post->update($request->all());
 
         if($request->file('file')){
@@ -135,7 +140,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        //
+        // Regla de Autorizacion
+        $this->authorize('author', $post);
         $post->delete();
         return redirect()->route('admin.posts.index');
     }
